feat:add clean code and solid in protected folder
